% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a, b}{are numeric scalars or integers.}
}
\value{
The greatest common divisor of \code{a} and \code{b}.
}
\description{
Finds the greatest common divisor following the euclidean algorithm.
}
\examples{
euclidean(1013253152, 1251514)
euclidean(85121215, 11182625192620156)

}
\references{
Wikipedia contributors. (2022, August 17). Euclidean algorithm. In Wikipedia, The Free Encyclopedia. Retrieved 11:30, September 9, 2022, from \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
