% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3Pack.R
\docType{package}
\name{JLPack}
\alias{JLPack}
\title{Euclidean & Dijkstra}
\description{
Euclidean computes the greatest commun divisor and the Dijkstra finds the shortest paths between nodes in a graph.
}
\author{
Leonard Persson Norblad \email{leope744@student.liu.se}

Jacob Sibanda \email{jacsi259@student.liu.se}
}
