% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph}
\format{
A \code{data.frame} of 18 rows and 3 colums.

\describe{ 
\item{v1}{First set of vertices}
\item{v2}{Second set of vertices}
\item{w}{Distance between vertices}
 }
}
\usage{
wiki_graph
}
\description{
A graph containing vertices and distances between them. Could be used in the dijkstra algorithm.
}
\examples{
data(wiki_graph)
}
\references{
{Wikipedia contributors. (2022, August 21). Dijkstra's algorithm. In Wikipedia, The Free Encyclopedia. Retrieved 07:10, September 12, 2022, from \url{https://en.wikipedia.org/w/index.php?title=Dijkstra\%27s_algorithm&oldid=1105704278}}
}
\keyword{datasets}
