\name{lab3pack-package}
\alias{lab3pack-package}
\alias{lab3pack}
\docType{package}
\title{
\packageTitle{lab3pack}
}
\description{
\packageDescription{lab3pack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab3pack}
\packageIndices{lab3pack}

}
\author{
\packageAuthor{lab3pack}

Maintainer: \packageMaintainer{lab3pack}
}

\keyword{ package }


