% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation_wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph for 1.1.2 * dijkstra()}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\usage{
wiki_graph
}
\description{
a data.frame of a graph with three variables
}
\details{
v1, v2: the edges of the graph; w: the weight of the edge
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\seealso{
https://en.wikipedia.org/wiki/Graph_(mathematics)
}
\keyword{datasets}
